Added state machine for prey and new input system
All the scripts have been manually moved over from my branch. They are set up so they should be ready to be used, but I haven't moved the actual movement code into them yet. Will be doing that now.

I also added the new input system to the project. This messes up the pre-made fps script that every unity project seems to come with. We aren't using it for anything so I just deleted it. We're currently having both input systems enabled, but it would probably be for the best if we moved everything over to the new one and disable the old one eventually.