Added hunter state machine
Just copied all the prey scripts and changed all the P_ and p_ to H_ and h_. Mainly just a placeholder for now, but now we have the base ready if we want to make some hunter movement.

As mentioned in the last commit: It would definitely be possible to turn the state machine for prey into more of a generic one and re-use it for hunter as well, but since we only have a grand total of two character types I'm pretty sure this'll be easier.