Greatly improved sliding on steep slopes
Also figured out what the requirements for the bug where we very rarely glitch through the ground while on a slope is. Haven't fixed it though.

It sometimes happens when you move up a slope and start sliding. But only if the emergency clipthrough method has been activated and the player hasn't changed their direction of movement too much. When they then start sliding the changing of the size of the capsule collider manages to just about allow them to clip through the ground. I have a couple of ideas for fixes in mind but it is also rare enough that fixing it is not a high prio.