Dash now works, no UI though
Also fiddled around with the camera and made the hitbox 50% larger to keep the animation within it better. That might mess up some collisions but I don't think it should be a mayor issue. Though I have not updated the multiplayer scenes with the new version of the player prefab from my scene yet.