Hopefully final prey version
Created what hopefully is the final version of the prey character. Removed a bunch of unused code to slim down the scripts a bit. Probably still a lot here that isn't used, but I only removed the stuff I know isn't used.
Removed mentions of the "ability manager" in the state machine because i moved that code into the state manager instead since we're only using a dash and double jump and don't need an entire extra system to manage just those two things.