Connected both state managers to interacter
also made them not inherit from the same class anymore since that's not needed. haven't deleted the "baseStateManager" class yet only because i don't want to accidentally overwrite anyone's potential work there. That should be considered useless for now and be deleted as soon as we're sure no one uses it for anything