Map moving works for all players at once. Prefab switcher turned back on.
With the way our system is built we cannot cleanly ensure that only the host can swap maps. So, we'll just pray that we don't have any mischeivous players. With that said, we must disable lobby joining during game, and if you leave a lobby you should just,, leave the game.