Hunter now playable in singleplayer
The hunter singleplayer map is just a copy of the prey one so far and does not have its intended functionality.
Also found a very easy way to clip out of basically any map. It's not likely to happen accidentally during gameplay but very easy to do when you know how it works.
Just spam slide while touching the roof with your head and double jumping and you go through it.