Added info text to the hunter Tutorial
Also fixed the error when missing an attack in the hunter tutorial, allowing people who aren't god gamers like myself to play.
Also also fixed the attack UI to work with singleplayer and remade it to be less resource intensive by removing the update function which wasn't needed to operate it.