Finished all the code for both Prey and Hunter trial
This includes game starting when you exit the start area, resetting the game, a timer which saves your time (should save it even after closing and reopening game), etc.
Still need to make actual maps for the trials still.
Also fixed a bunch of "yellow errors that has been in the editor for months now. If this breaks something tell me.